# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quantum-Resistant Blockchain API — CRYSTALS-Dilithium5, AES-256-GCM, SHA3-512
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterRequest {
  /**
   * @minLength 3
   * @maxLength 32
   */
  username: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export interface UserProfile {
  id: string;
  username: string;
  /** Dilithium5 public key (hex) */
  publicKey: string;
  /** Current token balance */
  balance: number;
  createdAt: string;
}

export interface AuthResponse {
  user: UserProfile;
  message: string;
}

export interface UsersListResponse {
  users: UserProfile[];
}

export interface BalanceResponse {
  publicKey: string;
  username: string;
  balance: number;
}

export interface SuccessMessage {
  message: string;
}

export interface Transaction {
  id: string;
  /** Sender public key (hex) */
  sender: string;
  senderUsername?: string;
  /** Recipient public key (hex) */
  recipient: string;
  recipientUsername?: string;
  amount: number;
  /** AES-256-GCM encrypted transaction metadata */
  encryptedData?: string;
  /** Dilithium5 signature (hex) */
  signature: string;
  timestamp: string;
  verified?: boolean;
}

export interface Block {
  index: number;
  timestamp: string;
  transactions: Transaction[];
  /** SHA3-512 hash of previous block */
  previousHash: string;
  /** SHA3-512 hash of this block */
  hash: string;
  nonce: number;
  minerAddress: string;
  minerUsername?: string;
}

export interface BlockchainResponse {
  chain: Block[];
  length: number;
  difficulty: number;
}

export interface ValidationResult {
  valid: boolean;
  message: string;
  checks: string[];
  invalidBlocks: number[];
}

export interface MineBlockRequest {
  /** Public key of miner (optional, uses logged-in user if omitted) */
  minerPublicKey?: string;
}

export interface SubmitTransactionRequest {
  /** Recipient's Dilithium5 public key */
  recipientPublicKey: string;
  /** Amount of tokens to transfer */
  amount: number;
  /** Optional transaction note (will be encrypted with AES-256-GCM) */
  note?: string;
}

export interface TransactionResponse {
  success: boolean;
  transactionId: string;
  message: string;
  newBalance: number;
}

export interface PendingTransactionsResponse {
  transactions: Transaction[];
  count: number;
}

export interface ErrorResponse {
  error: string;
  message: string;
}