# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Quantum-Resistant Blockchain API — CRYSTALS-Dilithium5, AES-256-GCM, SHA3-512
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: users
    description: User account operations
  - name: blockchain
    description: Blockchain operations
  - name: transactions
    description: Transaction operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user account
      description: Creates a new user, generates Dilithium5 key pair, grants 100 starting tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error or username taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Log out current session
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (for recipient selection)
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsersListResponse"

  /users/{publicKey}/balance:
    get:
      operationId: getUserBalance
      tags: [users]
      summary: Get balance for a user by public key
      parameters:
        - name: publicKey
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BalanceResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /blockchain:
    get:
      operationId: getBlockchain
      tags: [blockchain]
      summary: Get the full blockchain
      responses:
        "200":
          description: Blockchain data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlockchainResponse"

  /blockchain/validate:
    get:
      operationId: validateBlockchain
      tags: [blockchain]
      summary: Validate the blockchain integrity and all Dilithium5 signatures
      responses:
        "200":
          description: Validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationResult"

  /blockchain/mine:
    post:
      operationId: mineBlock
      tags: [blockchain]
      summary: Mine a new block (requires authentication)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MineBlockRequest"
      responses:
        "200":
          description: Block mined successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Block"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /transactions/pending:
    get:
      operationId: getPendingTransactions
      tags: [transactions]
      summary: Get pending transactions
      responses:
        "200":
          description: Pending transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PendingTransactionsResponse"

  /transactions/submit:
    post:
      operationId: submitTransaction
      tags: [transactions]
      summary: Submit a transaction (requires authentication)
      description: Signs and submits a transaction using the logged-in user's Dilithium5 keys
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitTransactionRequest"
      responses:
        "200":
          description: Transaction submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionResponse"
        "400":
          description: Insufficient balance or invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    RegisterRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 32
        password:
          type: string
          minLength: 6
      required: [username, password]

    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]

    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserProfile"
        message:
          type: string
      required: [user, message]

    UserProfile:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        publicKey:
          type: string
          description: Dilithium5 public key (hex)
        balance:
          type: number
          description: Current token balance
        createdAt:
          type: string
          format: date-time
      required: [id, username, publicKey, balance, createdAt]

    UsersListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/UserProfile"
      required: [users]

    BalanceResponse:
      type: object
      properties:
        publicKey:
          type: string
        username:
          type: string
        balance:
          type: number
      required: [publicKey, username, balance]

    SuccessMessage:
      type: object
      properties:
        message:
          type: string
      required: [message]

    Transaction:
      type: object
      properties:
        id:
          type: string
        sender:
          type: string
          description: Sender public key (hex)
        senderUsername:
          type: string
        recipient:
          type: string
          description: Recipient public key (hex)
        recipientUsername:
          type: string
        amount:
          type: number
        encryptedData:
          type: string
          description: AES-256-GCM encrypted transaction metadata
        signature:
          type: string
          description: Dilithium5 signature (hex)
        timestamp:
          type: string
          format: date-time
        verified:
          type: boolean
      required: [id, sender, recipient, amount, signature, timestamp]

    Block:
      type: object
      properties:
        index:
          type: integer
        timestamp:
          type: string
          format: date-time
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        previousHash:
          type: string
          description: SHA3-512 hash of previous block
        hash:
          type: string
          description: SHA3-512 hash of this block
        nonce:
          type: integer
        minerAddress:
          type: string
        minerUsername:
          type: string
      required: [index, timestamp, transactions, previousHash, hash, nonce, minerAddress]

    BlockchainResponse:
      type: object
      properties:
        chain:
          type: array
          items:
            $ref: "#/components/schemas/Block"
        length:
          type: integer
        difficulty:
          type: integer
      required: [chain, length, difficulty]

    ValidationResult:
      type: object
      properties:
        valid:
          type: boolean
        message:
          type: string
        checks:
          type: array
          items:
            type: string
        invalidBlocks:
          type: array
          items:
            type: integer
      required: [valid, message, checks, invalidBlocks]

    MineBlockRequest:
      type: object
      properties:
        minerPublicKey:
          type: string
          description: Public key of miner (optional, uses logged-in user if omitted)
      required: []

    SubmitTransactionRequest:
      type: object
      properties:
        recipientPublicKey:
          type: string
          description: Recipient's Dilithium5 public key
        amount:
          type: number
          description: Amount of tokens to transfer
        note:
          type: string
          description: Optional transaction note (will be encrypted with AES-256-GCM)
      required: [recipientPublicKey, amount]

    TransactionResponse:
      type: object
      properties:
        success:
          type: boolean
        transactionId:
          type: string
        message:
          type: string
        newBalance:
          type: number
      required: [success, transactionId, message, newBalance]

    PendingTransactionsResponse:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        count:
          type: integer
      required: [transactions, count]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error, message]