# lib/api-zod/src/generated/types/balanceResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quantum-Resistant Blockchain API — CRYSTALS-Dilithium5, AES-256-GCM, SHA3-512
 * OpenAPI spec version: 0.2.0
 */

export interface BalanceResponse {
  publicKey: string;
  username: string;
  balance: number;
}