# lib/api-zod/src/generated/types/block.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quantum-Resistant Blockchain API — CRYSTALS-Dilithium5, AES-256-GCM, SHA3-512
 * OpenAPI spec version: 0.2.0
 */
import type { Transaction } from "./transaction";

export interface Block {
  index: number;
  timestamp: Date;
  transactions: Transaction[];
  /** SHA3-512 hash of previous block */
  previousHash: string;
  /** SHA3-512 hash of this block */
  hash: string;
  nonce: number;
  minerAddress: string;
  minerUsername?: string;
}