# lib/api-zod/src/generated/types/blockchainResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quantum-Resistant Blockchain API — CRYSTALS-Dilithium5, AES-256-GCM, SHA3-512
 * OpenAPI spec version: 0.2.0
 */
import type { Block } from "./block";

export interface BlockchainResponse {
  chain: Block[];
  length: number;
  difficulty: number;
}