# lib/api-zod/src/generated/types/registerRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quantum-Resistant Blockchain API — CRYSTALS-Dilithium5, AES-256-GCM, SHA3-512
 * OpenAPI spec version: 0.2.0
 */

export interface RegisterRequest {
  /**
   * @minLength 3
   * @maxLength 32
   */
  username: string;
  /** @minLength 6 */
  password: string;
}