# lib/api-zod/src/generated/types/submitTransactionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quantum-Resistant Blockchain API — CRYSTALS-Dilithium5, AES-256-GCM, SHA3-512
 * OpenAPI spec version: 0.2.0
 */

export interface SubmitTransactionRequest {
  /** Recipient's Dilithium5 public key */
  recipientPublicKey: string;
  /** Amount of tokens to transfer */
  amount: number;
  /** Optional transaction note (will be encrypted with AES-256-GCM) */
  note?: string;
}